Moved task types to task interface and deleted types module
