+++ src/seb/interfaces/language.py
@@ -1,0 +1,4 @@
+from typing import Literal
+
+Language = Literal["da", "nb", "nn", "sv", "da-bornholm", "is", "fo"]
+languages_in_seb: list[Language] = ["da", "nb", "nn", "sv", "da-bornholm", "is", "fo"]
