+++ src/seb/registered_models/e5_mistral.py
@@ -4,13 +4,12 @@ from typing import Any, TypeVar
 
 import torch
 import torch.nn.functional as F
+from numpy.typing import ArrayLike
 from torch import Tensor
 from transformers import AutoModel, AutoTokenizer, BatchEncoding
 
 from seb import models
 from seb.interfaces.model import EmbeddingModel, Encoder, ModelMeta
-
-from ..types import ArrayLike
 
 T = TypeVar("T")
 
@@ -31,7 +30,9 @@ class E5Mistral(Encoder):
         self.load_model()
 
     def load_model(self):
-        self.tokenizer = AutoTokenizer.from_pretrained("intfloat/e5-mistral-7b-instruct")
+        self.tokenizer = AutoTokenizer.from_pretrained(
+            "intfloat/e5-mistral-7b-instruct"
+        )
         self.model = AutoModel.from_pretrained("intfloat/e5-mistral-7b-instruct")
 
     def preprocess(self, sentences: Sequence[str]) -> BatchEncoding:
@@ -52,7 +53,9 @@ class E5Mistral(Encoder):
             [*input_ids, self.tokenizer.eos_token_id]
             for input_ids in batch_dict["input_ids"]  # type: ignore
         ]
-        batch_dict = self.tokenizer.pad(batch_dict, padding=True, return_attention_mask=True, return_tensors="pt")
+        batch_dict = self.tokenizer.pad(
+            batch_dict, padding=True, return_attention_mask=True, return_tensors="pt"
+        )
 
         return batch_dict
